+++ neo/test/iotest/test_axonio.py
@@ -30,6 +30,15 @@ class TestAxonIO(BaseTestIO, unittest.TestCase):
     files_to_download = files_to_test
     ioclass = AxonIO
 
+    def test_read_protocol(self):
+        reader = AxonIO(filename=self.get_filename_path('File_axon_7.abf'))
+        segments = reader.read_protocol()
+    
+    def test_annotations(self):
+        reader = AxonIO(filename=self.get_filename_path('File_axon_2.abf'))
+        bl = reader.read_block()
+        ev = bl.segments[0].events[0]
+        assert 'comments' in ev.annotations
 
 if __name__ == "__main__":
     unittest.main()
